traceset_v2: Removed single value classes in favour of using 1-length arrays
Also split the code from ParameterUtils into TraceParameter and HexUtils
